Close connections if the local address is no longer valid.
When we detect that there are changes on the network interfaces, check for each active connection whether the local side of the connection has an address that exists on at least one network interface. If not, then communication via that connection is not possible. Instead of waiting for a timeout, immediately terminate those connections.