Fix potential segmentation fault on iOS
The PONG handler could call freeaddrinfo() on a struct that was not allocated with getaddrinfo(). On most platforms this apparently works fine, but on iOS it will try to free memory that wasn't allocated. Fix this by moving the code to reset an outgoing_t to a separate function, and calling that from the PONG handler.